Reorder logic, change data type, etc

+++ include/syn_scan.h
@@ -2,6 +2,6 @@
 #define SYN_SCAN_H
 
 int port_scan(char *address, unsigned short *port_arr, int count, int print_state);
-int *parse_ports(const char *port_str, int *port_count);
+unsigned short *parse_ports(const char *port_str, int *port_count);
 
 #endif
